Support multiple games.
This patch adds support for 'custom' games. A custom game applies
to a subset of the projects within the server, and scores for the
game are maintained independently of other games. Each game's
score can also be reset without disturbing other games.

Games can be configured on the main Jenkins config page, and the
leader board page allows you to pick which game's scores to view.

Regardless of if custom games are used, a default game still
tracks scores for all projects (maintaining the current behavior).
Care has been taken that the persistence of score data is backwards
compatible - old score information will still work with this update.